+++ src/main/java/it/gov/pagopa/register/dto/utils/EventDetails.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.register.utils;
+package it.gov.pagopa.register.dto.utils;
 
 import lombok.Getter;
